attempt_discovery once before thread
This should ensure that a connection has been found prior to a write
attempt. Without this, stringing a write to the end of Producer.new when
using lookupd's will likely result in `No connections available`

This could also take some time... meaning the discover_repeatedly method
call may take some time before it returns... This could be a downside.